Refactor time utility functions to a dedicated module for better organization and reuse
