+++ src/anonimizacion/seedwork/infraestructura/utils.py
@@ -7,19 +7,6 @@ epoch = datetime.datetime.utcfromtimestamp(0)
 def time_millis():
     return int(time.time() * 1000)
 
-def unix_time_millis(fecha):
-    """
-    Convierte una fecha en formato datetime a milisegundos desde epoch.
-    """
-    if fecha is None:
-        return 0
-    try:
-        epoch = datetime(1970, 1, 1)
-    except Exception as e:
-        print(e)
-        
-    return int((fecha - epoch).total_seconds() * 1000)
-
 def millis_a_datetime(millis):
     if isinstance(millis, datetime.datetime):
         return millis
