+++ src/validacion/seedwork/infraestructura/timeUtils.py
@@ -1,0 +1,6 @@
+import datetime
+
+epoch = datetime.datetime.utcfromtimestamp(0)
+
+def unix_time_millis(dt):
+    return (dt - epoch).total_seconds() * 1000.0
